+++ app/Models/admin/Repository.php
@@ -9,6 +9,12 @@ use Illuminate\Database\Eloquent\Model;
 class Repository extends Model
 {
     use HasFactory;
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $fillable = ['url', 'description'];
     public function user()
     {
         return $this->belongsTo(User::class);
