+++ database/factories/Admin/RepositoryFactory.php
@@ -18,7 +18,9 @@ class RepositoryFactory extends Factory
     public function definition()
     {
         return [
-            'user_id' =>User::factory()
+            'user_id'     => User::factory(),
+            'url'         => $this->faker->url,
+            'description' => $this->faker->text,
         ];
     }
 }
