--- tests/Feature/admin/RouteTest.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace Tests\Feature\admin;
-
-use Tests\TestCase;
-
-class RouteTest extends TestCase
-{
-    /**
-     * A basic feature test example.
-     *
-     * @return void
-     */
-    /********************************************
-     * Método para verificar si están logueados *
-     ********************************************/
-    public function test_guest()
-    {
-        // Redireccionamos a login
-        $this->get('repositories')->assertRedirect('login'); //index
-        $this->get('repositories/1')->assertRedirect('login'); //show
-        $this->get('repositories/1/edit')->assertRedirect('login'); //edit
-        $this->put('repositories/1')->assertRedirect('login'); //update
-        $this->delete('repositories/1')->assertRedirect('login'); //destroy
-        $this->get('repositories/create')->assertRedirect('login'); //create
-        $this->post('repositories', [])->assertRedirect('login'); //store - save
-    }
-}
